phân quyền mod

+++ app/Http/Controllers/Admin/BookCommentController.php
@@ -77,6 +77,10 @@ class BookCommentController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::check() || Auth::user()->role->name !== 'mod' && Auth::user()->role->name !== 'admin') {
+            return redirect()->route('bookComment.index')->with('error', 'Bạn không có quyền xóa bình luận!');
+        }
+
         $bookComment = $this->getbookCommentById($id);
         $bookComment->delete();
